Fixing #14 Incorrect application of activation with GEMM when reduce_k > 1

+++ src/kernels/activation.cl
@@ -1,9 +1,9 @@
 #include "defs.h"
 
 __kernel
-void activation(int size,__global dtype *a,ulong a_offset, __global dtype *c,ulong c_offset)
+void activation(ulong size,__global dtype *a,ulong a_offset, __global dtype *c,ulong c_offset)
 {
-    int pos = get_global_id(0);
+    ulong pos = get_global_id(0);
     if(pos >= size)
         return;
     a+=a_offset;
@@ -13,9 +13,9 @@ void activation(int size,__global dtype *a,ulong a_offset, __global dtype *c,ulo
 
 
 __kernel
-void activation_diff(int size,__global dtype *y,ulong y_offset, __global dtype *dy,ulong dy_offset,__global dtype *dx,ulong dx_offset,dtype beta)
+void activation_diff(ulong size,__global dtype *y,ulong y_offset, __global dtype *dy,ulong dy_offset,__global dtype *dx,ulong dx_offset,dtype beta)
 {
-    int pos = get_global_id(0);
+    ulong pos = get_global_id(0);
     if(pos >= size)
         return;
     y+=y_offset;
